+++ dotmotif/executors/NetworkXExecutor.py
@@ -14,9 +14,9 @@ See the License for the specific language governing permissions and
 limitations under the License.`
 """
 
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, List, Tuple
+import copy
 import networkx as nx
-import pandas as pd
 
 from .Executor import Executor
 
@@ -61,6 +61,27 @@ def _edge_satisfies_constraints(edge_attributes: dict, constraints: dict) -> boo
     return True
 
 
+def _edge_satisfies_many_constraints_for_muligraph_any_edges(edge_attributes: dict, constraints: dict) -> List[Tuple[str, str, str]]:
+    """
+    Returns a subset of constraints that this edge matches, in the form (key, op, val).
+    """
+    matched_constraints = []
+    for key, clist in constraints.items():
+        for operator, values in clist.items():
+            for value in values:
+                keyvalue_or_none = edge_attributes.get(key, None)
+                try:
+                    operator_success = _OPERATORS[operator](keyvalue_or_none, value)
+                except TypeError:
+                    # If you encounter a type error, that means the comparison
+                    # could not possibly succeed
+                    # # TODO: unless you tried a comparison
+                    # against an undefined value (i.e. VALUE != undefined)
+                    operator_success = False
+                if operator_success:
+                    matched_constraints.append((key, operator, value))
+    return matched_constraints
+
 def _node_satisfies_constraints(node_attributes: dict, constraints: dict) -> bool:
     """
     Check if a single node satisfies the constraints.
@@ -89,6 +110,11 @@ class NetworkXExecutor(Executor):
 
         Arguments:
             graph (networkx.Graph)
+            multigraph_edge_match (str: 'any'): A string ('any' or 'all') that
+                determines how to match edges between nodes in the graph. If
+                'any', then any edge between nodes can match the constraints
+                to satisfy the motif. If 'all', then all edges between nodes
+                must match the constraints to satisfy the motif.
 
         Returns:
             None
@@ -100,6 +126,15 @@ class NetworkXExecutor(Executor):
             raise ValueError(
                 "You must pass a graph to the NetworkXExecutor constructor."
             )
+
+        # Allow the user to set whether ALL edges should match when considering
+        # a multigraph or if ANY edges should be considered when matching.
+        # We only do this if the graph is a multigraph.
+        self._host_is_multigraph = False
+        if self.graph.is_multigraph():
+            self._host_is_multigraph = True
+            self._multigraph_edge_match = kwargs.get("multigraph_edge_match", "any")
+            assert self._multigraph_edge_match in ("all", "any"), "_multigraph_edge_match must be one of 'all' or 'any'."
 
     def _validate_node_constraints(
         self, node_isomorphism_map: dict, graph: nx.DiGraph, constraints: dict
@@ -200,7 +235,75 @@ class NetworkXExecutor(Executor):
                 return False
         return True
 
-    def count(self, motif: "dotmotif", limit: int = None):
+    def _validate_multigraph_all_edge_constraints(
+        self, node_isomorphism_map: dict, graph: nx.DiGraph, constraints: dict
+    ):
+        """
+        Reuses logic from the simple _validate_edge_constraints case.
+
+        Sole modification is that in the multigraph case, ALL edges must match
+        for an edge to be considered valid. If ANY of the edges between two
+        nodes mismatch the constraints, the mapping fails.
+
+        """
+        for (motif_U, motif_V), constraint_list in constraints.items():
+            # Get graph nodes (from this isomorphism)
+            graph_u = node_isomorphism_map[motif_U]
+            graph_v = node_isomorphism_map[motif_V]
+
+            # Check each edge in graph for constraints
+            for _, _, edge_attrs in graph.edges((graph_u, graph_v), data=True):
+                if not _edge_satisfies_constraints(edge_attrs, constraint_list):
+                    # Fail fast
+                    return False
+        return True
+
+    def _validate_multigraph_any_edge_constraints(
+        self, node_isomorphism_map: dict, graph: nx.DiGraph, constraints: dict
+    ):
+        """
+        Reuses logic from the simple _validate_edge_constraints case.
+
+        Sole modification is that in the multigraph case, ANY edge can match
+        for an edge to be considered valid. If ANY of the edges between two
+        nodes match the constraints, the mapping succeeds.
+
+        """
+        for (motif_U, motif_V), constraint_list in constraints.items():
+            # Get graph nodes (from this isomorphism)
+            graph_u = node_isomorphism_map[motif_U]
+            graph_v = node_isomorphism_map[motif_V]
+
+            # check each edge in the graph for the constraints.
+            # if you find an edge that matches, REMOVE that constraint from
+            # the list and continue checking.
+            # if you get to the end of the list of edges and there are any
+            # constrains left, the mapping fails.
+
+            # Check each edge in graph for constraints
+            constraint_list_copy = copy.deepcopy(constraint_list)
+            for _, _, edge_attrs in graph.edges((graph_u, graph_v), data=True):
+                matched_constraints = _edge_satisfies_many_constraints_for_muligraph_any_edges(edge_attrs, constraint_list_copy)
+                if matched_constraints:
+                    # Remove matched constraints from the list
+                    for constraint in matched_constraints:
+                        (key, operator, value) = constraint
+                        # remove `value` from the list of [key][operator].
+                        # if the list is empty, remove the entire key.
+                        constraint_list_copy[key][operator].remove(value)
+                        if len(constraint_list_copy[key][operator]) == 0:
+                            del constraint_list_copy[key][operator]
+                        if not constraint_list_copy[key]:
+                            del constraint_list_copy[key]
+
+            # if there are any constraints left over, the mapping failed
+            if len(constraint_list_copy) > 0:
+                return False
+
+        return True
+
+
+    def count(self, motif: "dotmotif.Motif", limit: int = None):
         """
         Count the occurrences of a motif in a graph.
 
@@ -208,12 +311,12 @@ class NetworkXExecutor(Executor):
         """
         return len(self.find(motif, limit))
 
-    def find(self, motif: "dotmotif", limit: int = None):
+    def find(self, motif: "dotmotif.Motif", limit: int = None):
         """
         Find a motif in a larger graph.
 
         Arguments:
-            motif (dotmotif.dotmotif)
+            motif (dotmotif.Motif)
 
         """
         # TODO: Can add constraints on iso node assignment. If we do this a
@@ -263,12 +366,19 @@ class NetworkXExecutor(Executor):
             if _doesnt_have_any_of_motifs_negative_edges(mapping)
         ]
 
+        _edge_constraint_validator = (
+            self._validate_edge_constraints if not self._host_is_multigraph else (
+                self._validate_multigraph_all_edge_constraints
+                if self._multigraph_edge_match == "all"
+                else self._validate_multigraph_any_edge_constraints
+            )
+        )
         # Now, filter on attributes:
         res = [
             r
             for r in results
             if (
-                self._validate_edge_constraints(
+                _edge_constraint_validator(
                     r, self.graph, motif.list_edge_constraints()
                 )
                 and self._validate_node_constraints(